cors 설정

+++ src/main/java/com/hi_e/vacation/repository/VacationRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hi_e.vacation.entity.Vacation;
 
-public interface VacationRepository extends JpaRepository<Vacation, Integer> {
-	
+public interface VacationRepository extends JpaRepository<Vacation, Long> {
 }
